requirements.txt
Django==1.6
Markdown==2.3.1
django-filter==0.7
django-picklefield==0.3.1
djangorestframework==2.3.9
numpy==1.8.0
pandas==0.12.0
psycopg2==2.5.1
python-dateutil==2.2
pytz==2013.8
scipy==0.13.1
six==1.4.1
wsgiref==0.1.2
